Fix sanitizer read error. Add ability to specify build configuration for windows tests

+++ EnvironmentSimulator/Libraries/esminiLib/esminiLib.cpp
@@ -121,7 +121,7 @@ static void ConvertArguments()
   for (unsigned int i = 0; i < static_cast<unsigned int>(argc_); i++)
   {
     argv_[i] = reinterpret_cast<char *>(malloc((args_v[i].size() + 1) * sizeof(char)));
-    StrCopy(argv_[i], args_v[i].c_str(), static_cast<unsigned int>(args_v[i].size()) + 1);
+    StrCopy(argv_[i], args_v[i].c_str(), args_v[i].size() + 1);
     argument_list += std::string(" ") + argv_[i];
   }
   LOG("Player arguments: %s", argument_list.c_str());
